requirements.txt
numpy
h5py
tensorflow-gpu
jax==0.2.6
objax==1.2.0
seaborn>=0.11
